+++ src/main/java/asc/portfolio/ascSb/web/dto/ticket/TicketResponseDto.java
@@ -12,15 +12,8 @@ public class TicketResponseDto {
 
     private TicketStateType isValidTicket;
     private LocalDateTime fixedTermTicket; // 기간제 티켓 날짜 => fixedTermTicket - createDate 시간으로 남은기간 계산
-    private Integer partTimeTicket; // 결제한 시간제 티켓시간 // 50시간, 100시간
-    private Integer remainingTime; // 시간제 티켓 남은시간
+    private Long partTimeTicket; // 결제한 시간제 티켓시간 // 50시간, 100시간
+    private Long remainingTime; // 시간제 티켓 남은시간
 
     private long period;
-
-//    public TicketResponseDto(TicketResponseDto entity) {
-//        this.isValidTicket = entity.getIsValidTicket();
-//        this.fixedTermTicket = entity.getFixedTermTicket();
-//        this.partTimeTicket = entity.getPartTimeTicket();
-//        this.remainingTime = entity.getRemainingTime();
-//    }
 }
